+++ net35/CountlySample/Program.cs
@@ -145,7 +145,7 @@ namespace CountlySample
                 }
             };
 
-            await Countly.EndSession();
+            await Countly.Instance.SessionEnd();
         }