+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 20.11.0
 * Fixed bug that occurred while getting device and system information on restricted machines.
+* Removed deprecated function "EndSession"
+* Removed deprecated function "StartSession"
+* [net35, net40] Removed deprecated function "SetCustomDataPath"
 
 20.05.1
 * Fixed bug that occured when no cpu_id was got returned from OS